tests/wpt/web-platform-tests/websockets/Create-asciiSep-protocol-string.htm

<!DOCTYPE html>
<html>
<head>
<title>W3C WebSocket API - Create WebSocket - ascii protocol string with separator</title>
<script type="text/javascript" src="/resources/testharness.js"></script>
<script type="text/javascript" src="/resources/testharnessreport.js"></script>
<script type="text/javascript" src="websocket.js?pipe=sub"></script>
</head>
<body>
<div id="log"></div>
<script type="text/javascript">
test(function () {
  assert_throws("SYNTAX_ERR", function () {
    var asciiWithSep = "/echo";
    CreateWebSocketWithAsciiSep(asciiWithSep)
  });
}, "W3C WebSocket API - Create WebSocket - Pass a valid URL and a protocol string with an ascii separator character - SYNTAX_ERR is thrown")
</script>
</body>
</html>
